Rename SelfieSegmentator to SelfieSegmenter
Renamed the directory and header file from 'SelfieSegmentator' to 'SelfieSegmenter' and updated all relevant include paths. This improves naming consistency across the codebase.
