+++ src/Core/RenderingContext.hpp
@@ -31,7 +31,7 @@ with this program. If not, see <https://www.gnu.org/licenses/>
 
 #include "MainEffect.hpp"
 #include "Preset.hpp"
-#include "../SelfieSegmentator/SelfieSegmenter.hpp"
+#include "../SelfieSegmenter/SelfieSegmenter.hpp"
 
 namespace KaitoTokyo {
 namespace BackgroundRemovalLite {
